% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_contact_point}
\alias{create_contact_point}
\title{Create a new contact point of a supported type (among: "email", "slack").}
\usage{
create_contact_point(
  project_id,
  type,
  name,
  addresses = NULL,
  webhook_url = NULL
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{type}{contact point type among "email" or "slack".}

\item{name}{contact point name.}

\item{addresses}{contact point addresses.}

\item{webhook_url}{contact point webhook_url.}
}
\value{
list - parsed content of the contact point.
}
\description{
Create a new contact point of a supported type (among: "email", "slack").
}
