% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset_embedding}
\alias{create_dataset_embedding}
\title{Create a dataset embedding from a dataset_id.}
\usage{
create_dataset_embedding(dataset_id)
}
\arguments{
\item{dataset_id}{dataset id.}
}
\description{
Create a dataset embedding from a dataset_id.
}
