% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset_from_dataframe}
\alias{create_dataset_from_dataframe}
\title{Upload dataset from data frame.}
\usage{
create_dataset_from_dataframe(dataset_name, dataframe, zip = F)
}
\arguments{
\item{dataset_name}{given name of the dataset on the platform.}

\item{dataframe}{data.frame to upload.}

\item{zip}{is the temp file zipped before sending it to Prevision.io (default = F).}
}
\value{
parsed content of the dataset.
}
\description{
Upload dataset from data frame.
}
