% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset_from_datasource}
\alias{create_dataset_from_datasource}
\title{Create a dataset from an existing datasource.}
\usage{
create_dataset_from_datasource(project_id, dataset_name, datasource_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{dataset_name}{given name of the dataset on the platform.}

\item{datasource_id}{datasource id.}
}
\value{
parsed content of the dataset.
}
\description{
Create a dataset from an existing datasource.
}
