% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset_from_file}
\alias{create_dataset_from_file}
\title{Upload dataset from file name.}
\usage{
create_dataset_from_file(dataset_name, file)
}
\arguments{
\item{dataset_name}{given name of the dataset on the platform.}

\item{file}{path to the dataset.}
}
\value{
parsed content of the dataset.
}
\description{
Upload dataset from file name.
}
