% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{create_deployment_api_key}
\alias{create_deployment_api_key}
\title{Create a new API key for a deployed model.}
\usage{
create_deployment_api_key(deployment_id)
}
\arguments{
\item{deployment_id}{id of the deployment to create an API key on, can be obtained with get_deployments().}
}
\value{
list - API key information.
}
\description{
Create a new API key for a deployed model.
}
