% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{create_deployment_model}
\alias{create_deployment_model}
\title{Create a new deployment for a model.}
\usage{
create_deployment_model(
  project_id,
  name,
  experiment_id,
  main_model_experiment_version_id,
  challenger_model_experiment_version_id = NULL,
  access_type = c("fine_grained", "private", "public"),
  description = NULL,
  main_model_id,
  challenger_model_id = NULL
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{name}{name of the deployment.}

\item{experiment_id}{id of the experiment to deploy, can be obtained with get_experiment_id_from_name().}

\item{main_model_experiment_version_id}{id of the experiment_version to deploy, can be obtained with get_experiment_version_id().}

\item{challenger_model_experiment_version_id}{id of the challenger experiment_version to deploy, can be obtained with get_experiment_version_id().}

\item{access_type}{type of access of the deployment among "fine_grained" (project defined, default), "private" (instance) or "public" (everyone).}

\item{description}{description of the deployment.}

\item{main_model_id}{id of the model to deploy}

\item{challenger_model_id}{id of the challenger model to deploy}
}
\value{
list - parsed content of the deployment.
}
\description{
Create a new deployment for a model.
}
