% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{create_deployment_predictions}
\alias{create_deployment_predictions}
\title{Create predictions on a deployed model using a dataset.}
\usage{
create_deployment_predictions(deployment_id, dataset_id)
}
\arguments{
\item{deployment_id}{id of the deployment, can be obtained with get_deployments().}

\item{dataset_id}{id of the dataset to predict, can be obtained with get_dataset_id_from_name().}
}
\value{
integer - 200 on success.
}
\description{
Create predictions on a deployed model using a dataset.
}
