% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a new experiment on the platform.}
\usage{
create_experiment(project_id, name, provider, data_type, training_type)
}
\arguments{
\item{project_id}{id of the project in which we create the experiment.}

\item{name}{name of the experiment.}

\item{provider}{provider of the experiment ("prevision-auto-ml" or "external")}

\item{data_type}{type of data ("tabular", "images" or "timeseries").}

\item{training_type}{type of the training you want to achieve ("regression", "classification", "multiclassification", "clustering", "object-detection" or "text-similarity").}
}
\value{
experiment information.
}
\description{
Create a new experiment on the platform.
}
