% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{create_export}
\alias{create_export}
\title{Export data using an existing exporter and the resource to export}
\usage{
create_export(exporter_id, type, dataset_id = NULL, prediction_id = NULL)
}
\arguments{
\item{exporter_id}{id of the exporter, can be obtained with get_exporters().}

\item{type}{type of data to export among \"dataset"\, \"validation-prediction\" or \"deployment-prediction\"}

\item{dataset_id}{id of the dataset to export (only for type == \"dataset\")}

\item{prediction_id}{id of the prediction to export (only for type == \"validation_prediction\" or type == \"deployment-prediction\")}
}
\value{
list - parsed content of the export.
}
\description{
Export data using an existing exporter and the resource to export
}
