% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{create_exporter}
\alias{create_exporter}
\title{Create a new exporter}
\usage{
create_exporter(
  project_id,
  connector_id,
  name,
  description = "",
  filepath = "",
  file_write_mode = "timestamp",
  database = "",
  table = "",
  database_write_mode = "append",
  bucket = ""
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{connector_id}{connector_id linked to the exporter.}

\item{name}{exporter name.}

\item{description}{description of the exporter.}

\item{filepath}{exporter path (for SFTP & FTP connector).}

\item{file_write_mode}{writing type when exporting a file (for SFT & FTP connector, among \"timestamp\", \"safe\" or \"replace\")}

\item{database}{exporter database (for SQL connector).}

\item{table}{exporter table (for SQL connector).}

\item{database_write_mode}{writing type when exporting data within a database (for SQL connector, among \"append\" or \"replace\").}

\item{bucket}{exporter bucket (for S3 connector).}
}
\value{
list - parsed content of the exporter.
}
\description{
Create a new exporter
}
