% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{create_pipeline}
\alias{create_pipeline}
\title{[BETA] Create a new pipeline of a supported type among "component", "template", "run".}
\usage{
create_pipeline(
  project_id,
  type,
  name,
  git_url = NULL,
  git_branch = NULL,
  repository_name = NULL,
  broker = NULL,
  config_dataset_id = NULL,
  nodes = NULL,
  pipeline_template_id = NULL,
  pipeline_parameters = NULL
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{type}{type of the pipeline to be retrieved among "component", "template", "run".}

\item{name}{name of the pipeline.}

\item{git_url}{url of the git repository than contains the component.}

\item{git_branch}{branch of the git repository than contains the component.}

\item{repository_name}{name of the git repository that contains the component.}

\item{broker}{broker of the git repository that contains the component.}

\item{config_dataset_id}{only for templates.}

\item{nodes}{list, only for templates.}

\item{pipeline_template_id}{id of the pipeline template to add for a run.}

\item{pipeline_parameters}{list of parameters for the run.}
}
\value{
list - parsed content of the pipeline.
}
\description{
[BETA] Create a new pipeline of a supported type among "component", "template", "run".
}
