% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{create_pipeline_trigger}
\alias{create_pipeline_trigger}
\title{[BETA] Trigger an existing pipeline run.}
\usage{
create_pipeline_trigger(pipeline_id)
}
\arguments{
\item{pipeline_id}{id of the pipeline run to trigger, can be obtained with get_pipelines().}
}
\value{
integer - 200 on success.
}
\description{
[BETA] Trigger an existing pipeline run.
}
