% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project_user}
\alias{create_project_user}
\title{Add user in and existing project.}
\usage{
create_project_user(project_id, user_mail, user_role)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{user_mail}{email of the user to be add.}

\item{user_role}{role to grand to the user among "admin", "contributor", "viewer" or "end_user".}
}
\value{
list - information of project's users.
}
\description{
Add user in and existing project.
}
