% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{createApp}
\alias{createApp}
\title{Create a new app}
\usage{
createApp(type, name, tag, route, modelId = NULL, path = NULL)
}
\arguments{
\item{type}{app type (can be "model", "notebook", "app")}

\item{name}{app name.}

\item{tag}{tag of the app (integer).}

\item{route}{app route.}

\item{modelId}{id of the model to deploy if you have selected "model" as type.}

\item{path}{path of your code if you have selected "notebook" or "app" as type.}
}
\value{
parsed content of the app.
}
\description{
Create a new app
}
