% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{createConnector}
\alias{createConnector}
\title{Create a new connector of a supported type (among: "SQL", "HIVE", "FTP", "SFTP", "S3", "GCP")}
\usage{
createConnector(
  type,
  name,
  host,
  port,
  username,
  password,
  googleCredentials = NULL
)
}
\arguments{
\item{type}{connector type.}

\item{name}{connector name.}

\item{host}{connector host.}

\item{port}{connector port.}

\item{username}{connector username.}

\item{password}{connector password.}

\item{googleCredentials}{google credentials JSON (for GCP only).}
}
\value{
parsed content of the connector.
}
\description{
Create a new connector of a supported type (among: "SQL", "HIVE", "FTP", "SFTP", "S3", "GCP")
}
