% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDataframeFromDataset}
\alias{createDataframeFromDataset}
\title{Create a dataframe from a datasetId.}
\usage{
createDataframeFromDataset(datasetId)
}
\arguments{
\item{datasetId}{dataset id.}
}
\value{
a R dataframe
}
\description{
Create a dataframe from a datasetId.
}
