% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDatasetFromDataframe}
\alias{createDatasetFromDataframe}
\title{Upload dataset from data frame.}
\usage{
createDatasetFromDataframe(datasetName, dataframe, zip = F)
}
\arguments{
\item{datasetName}{given name of the dataset on the platform.}

\item{dataframe}{data.frame to upload.}

\item{zip}{is the temp file zipped before sending it to Prevision.io?}
}
\value{
parsed content of the dataset object
}
\description{
Upload dataset from data frame.
}
