% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDatasetFromDatasource}
\alias{createDatasetFromDatasource}
\title{Create a dataset from an existing datasource.}
\usage{
createDatasetFromDatasource(datasetName, datasourceId)
}
\arguments{
\item{datasetName}{given name of the dataset on the platform.}

\item{datasourceId}{datasource id.}
}
\value{
parsed content of the dataset object
}
\description{
Create a dataset from an existing datasource.
}
