% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDatasetFromFilename}
\alias{createDatasetFromFilename}
\title{Upload dataset from file name.}
\usage{
createDatasetFromFilename(datasetName, local_path)
}
\arguments{
\item{datasetName}{given name of the dataset on the platform.}

\item{local_path}{path to the dataset.}
}
\value{
parsed content of the dataset.
}
\description{
Upload dataset from file name.
}
