% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{createDatasource}
\alias{createDatasource}
\title{Create a new datasource}
\usage{
createDatasource(
  connectorId,
  name,
  path = "",
  database = "",
  table = "",
  bucket = "",
  request = ""
)
}
\arguments{
\item{connectorId}{connectorId linked to the datasource.}

\item{name}{datasource name.}

\item{path}{datasource path (for SFTP & FTP connector).}

\item{database}{datasource database (for SQL & HIVE & HBASE connector).}

\item{table}{datasource table (for SQL & HIVE & HBASE connector).}

\item{bucket}{datasource bucket (for S3 connector).}

\item{request}{datasource request (for SQL & HIVE connector).}
}
\value{
parsed content of the datasource
}
\description{
Create a new datasource
}
