% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{createFolder}
\alias{createFolder}
\title{Upload folder from a local file.}
\usage{
createFolder(folder_name, local_path)
}
\arguments{
\item{folder_name}{given name of the folder on the platform.}

\item{local_path}{path to the folder.}
}
\value{
parsed content of the folder.
}
\description{
Upload folder from a local file.
}
