% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{delete_connector}
\alias{delete_connector}
\title{Delete an existing connector.}
\usage{
delete_connector(connector_id)
}
\arguments{
\item{connector_id}{id of the connector to be deleted, can be obtained with get_connectors().}
}
\description{
Delete an existing connector.
}
