% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{delete_contact_point}
\alias{delete_contact_point}
\title{Delete an existing contact_point}
\usage{
delete_contact_point(contact_point_id)
}
\arguments{
\item{contact_point_id}{id of the contact point to be deleted, can be obtained with get_contact_points().}
}
\value{
integer - 204 on success.
}
\description{
Delete an existing contact_point
}
