% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete an existing dataset.}
\usage{
delete_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{id of the dataset, can be obtained with get_datasets().}
}
\value{
integer - 204 on success.
}
\description{
Delete an existing dataset.
}
