% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{delete_datasource}
\alias{delete_datasource}
\title{Delete a datasource}
\usage{
delete_datasource(datasource_id)
}
\arguments{
\item{datasource_id}{id of the connector to be deleted, can be obtained with listConnectors().}
}
\description{
Delete a datasource
}
