% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{delete_deployment}
\alias{delete_deployment}
\title{Delete an existing deployment}
\usage{
delete_deployment(deployment_id)
}
\arguments{
\item{deployment_id}{id of the deployment, can be obtained with get_deployments().}
}
\value{
integer - 204 on success.
}
\description{
Delete an existing deployment
}
