% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{delete_experiment}
\alias{delete_experiment}
\title{Delete a experiment on the platform.}
\usage{
delete_experiment(experiment_id)
}
\arguments{
\item{experiment_id}{id of the experiment, can be obtained with get_experiments().}
}
\value{
integer - 204 on success.
}
\description{
Delete a experiment on the platform.
}
