% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{delete_exporter}
\alias{delete_exporter}
\title{Delete an exporter}
\usage{
delete_exporter(exporter_id)
}
\arguments{
\item{exporter_id}{id of the exporter to be deleted, can be obtained with get_exporters().}
}
\value{
integer - 204 on success.
}
\description{
Delete an exporter
}
