% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{delete_folder}
\alias{delete_folder}
\title{Delete an existing folder.}
\usage{
delete_folder(folder_id)
}
\arguments{
\item{folder_id}{id of the folder to be deleted.}
}
\description{
Delete an existing folder.
}
