% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{delete_project}
\alias{delete_project}
\title{Delete an existing project.}
\usage{
delete_project(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
200 on success
}
\description{
Delete an existing project.
}
