% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{delete_project_user}
\alias{delete_project_user}
\title{Delete user in and existing project.}
\usage{
delete_project_user(project_id, user_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{user_id}{user_id of the user to be delete, can be obtained with get_project_users().}
}
\value{
200 on success
}
\description{
Delete user in and existing project.
}
