% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{delete_scheduler}
\alias{delete_scheduler}
\title{Delete a scheduler task.}
\usage{
delete_scheduler(scheduler_id)
}
\arguments{
\item{scheduler_id}{id of the scheduler task to be deleted, can be obtained with get_scheduler().}
}
\value{
response status code (200 on success)
}
\description{
Delete a scheduler task.
}
