% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{delete_usecase_version}
\alias{delete_usecase_version}
\title{Delete a usecase_version on the platform.}
\usage{
delete_usecase_version(usecase_version_id)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}
}
\description{
Delete a usecase_version on the platform.
}
