% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{deleteApp}
\alias{deleteApp}
\title{Delete an app}
\usage{
deleteApp(appId)
}
\arguments{
\item{appId}{id of the app to be deleted, can be obtained with listApps().}
}
\value{
response status code (200 on success)
}
\description{
Delete an app
}
