% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{deleteConnector}
\alias{deleteConnector}
\title{Delete a connector.}
\usage{
deleteConnector(connectorId)
}
\arguments{
\item{connectorId}{id of the connector to be deleted, can be obtained with getConnectors().}
}
\description{
Delete a connector.
}
