% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete a dataset on the platform.}
\usage{
deleteDataset(datasetId)
}
\arguments{
\item{datasetId}{id of the dataset, can be obtained with getDatasets().}
}
\description{
Delete a dataset on the platform.
}
