% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{deleteDatasource}
\alias{deleteDatasource}
\title{Delete a datasource}
\usage{
deleteDatasource(datasourceId)
}
\arguments{
\item{datasourceId}{id of the connector to be deleted, can be obtained with listConnectors().}
}
\description{
Delete a datasource
}
