% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{deleteScheduler}
\alias{deleteScheduler}
\title{Delete a scheduler task}
\usage{
deleteScheduler(schedulerId)
}
\arguments{
\item{schedulerId}{id of the scheduler task to be deleted, can be obtained with listScheduler().}
}
\value{
response status code (200 on success)
}
\description{
Delete a scheduler task
}
