% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download a dataset from a dataset_id and write it in a folder.}
\usage{
download_dataset(dataset_id, path = getwd(), is_folder = FALSE)
}
\arguments{
\item{dataset_id}{dataset id.}

\item{path}{path (without / at the end) were to write the dataset.}

\item{is_folder}{TRUE if it's a folder dataset, FALSE (by default) otherwise.}
}
\value{
the complete path to the file written
}
\description{
Download a dataset from a dataset_id and write it in a folder.
}
