% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download a dataset from a datasetId and write it in a folder.}
\usage{
downloadDataset(datasetId, path = getwd(), isFolder = FALSE)
}
\arguments{
\item{datasetId}{dataset id.}

\item{path}{path (without / at the end) were to write the dataset}

\item{isFolder}{boolean. TRUE if it's a folder dataset, FALSE (by default) otherwise.}
}
\value{
the complete path to the file written
}
\description{
Download a dataset from a datasetId and write it in a folder.
}
