% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_best_model_id}
\alias{get_best_model_id}
\title{Get the model_id that provide the best predictive performance given usecase_id and version_number. If include_blend is false, it will return the model_id from the best "non blended" model.}
\usage{
get_best_model_id(usecase_id, version_number = 1, include_blend = TRUE)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{version_number}{number of the version of the usecase. 1 by default.}

\item{include_blend}{boolean, indicating if you want to retrieve the best model among blended models too.}
}
\value{
model_id.
}
\description{
Get the model_id that provide the best predictive performance given usecase_id and version_number. If include_blend is false, it will return the model_id from the best "non blended" model.
}
