% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{get_connector_info}
\alias{get_connector_info}
\title{Get information about connector from its id.}
\usage{
get_connector_info(connector_id)
}
\arguments{
\item{connector_id}{id of the connector to be retrieved, can be obtained with get_connectors().}
}
\value{
parsed content of the connector.
}
\description{
Get information about connector from its id.
}
