% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{get_connectors}
\alias{get_connectors}
\title{Get information of all connectors available.}
\usage{
get_connectors()
}
\value{
parsed content of all connectors.
}
\description{
Get information of all connectors available.
}
