% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_contact_point_info}
\alias{get_contact_point_info}
\title{Get a contact point information from its contact_point_id.}
\usage{
get_contact_point_info(contact_point_id)
}
\arguments{
\item{contact_point_id}{id of the contact point, can be obtained with get_contact_points().}
}
\value{
list - information of the contact point.
}
\description{
Get a contact point information from its contact_point_id.
}
