% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_contact_points}
\alias{get_contact_points}
\title{Get information of all contact points available for a given project_id.}
\usage{
get_contact_points(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
list - parsed content of all contact points for the supplied project_id.
}
\description{
Get information of all contact points available for a given project_id.
}
