% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_embedding}
\alias{get_dataset_embedding}
\title{Get a dataset embedding from a dataset_id.}
\usage{
get_dataset_embedding(dataset_id)
}
\arguments{
\item{dataset_id}{dataset id.}
}
\value{
integer - 200 on success.
}
\description{
Get a dataset embedding from a dataset_id.
}
