% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_head}
\alias{get_dataset_head}
\title{Show the head of a dataset from its id.}
\usage{
get_dataset_head(dataset_id)
}
\arguments{
\item{dataset_id}{id of the dataset, can be obtained with get_datasets().}
}
\value{
data.frame - head of the dataset.
}
\description{
Show the head of a dataset from its id.
}
