% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get a dataset from its id.}
\usage{
get_dataset_info(dataset_id)
}
\arguments{
\item{dataset_id}{id of the dataset, can be obtained with get_datasets().}
}
\value{
list - parsed content of the dataset.
}
\description{
Get a dataset from its id.
}
