% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_datasets}
\alias{get_datasets}
\title{Get information of all tabular datasets availables.}
\usage{
get_datasets()
}
\value{
parsed content of all datasets.
}
\description{
Get information of all tabular datasets availables.
}
