% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{get_datasource_id_from_name}
\alias{get_datasource_id_from_name}
\title{Get a datasource_id from a datasource_name If duplicated name, the first datasource_id that match it is retrieved}
\usage{
get_datasource_id_from_name(project_id, datasource_name)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects(project_id).}

\item{datasource_name}{name of the connector we are searching its id from. Can be obtained with get_datasources().}
}
\value{
id of the datasource if found.
}
\description{
Get a datasource_id from a datasource_name If duplicated name, the first datasource_id that match it is retrieved
}
