% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{get_datasource_info}
\alias{get_datasource_info}
\title{Get a datasource from its id.}
\usage{
get_datasource_info(datasource_id)
}
\arguments{
\item{datasource_id}{id of the data_sources to be retrieved, can be obtained with get_datasources().}
}
\value{
list - parsed content of the data_sources.
}
\description{
Get a datasource from its id.
}
