% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment-alert.R
\name{get_deployment_alert_id_from_name}
\alias{get_deployment_alert_id_from_name}
\title{Get a deployment_alert_id from a name and type for a given deployment_id.}
\usage{
get_deployment_alert_id_from_name(deployment_id, name)
}
\arguments{
\item{deployment_id}{id of the deployment, can be obtained with get_deployments().}

\item{name}{name of the deployment_alert we are searching its id from.}
}
\value{
character - id of the deployment_alert if found.
}
\description{
Get a deployment_alert_id from a name and type for a given deployment_id.
}
