% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment-alert.R
\name{get_deployment_alert_info}
\alias{get_deployment_alert_info}
\title{Get information about a deployment_alert for a given deployed model.}
\usage{
get_deployment_alert_info(deployment_id, deployment_alert_id)
}
\arguments{
\item{deployment_id}{id of the deployment, can be obtained with get_deployments().}

\item{deployment_alert_id}{id of the deployment_alert to be retrieved, can be obtained with get_deployment_alerts().}
}
\value{
list - parsed content of the deployment_alert.
}
\description{
Get information about a deployment_alert for a given deployed model.
}
