% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment-alert.R
\name{get_deployment_alerts}
\alias{get_deployment_alerts}
\title{Get information of all alerts related to a deployment_id.}
\usage{
get_deployment_alerts(deployment_id)
}
\arguments{
\item{deployment_id}{id of the project, can be obtained with get_deployments().}
}
\value{
list - parsed content of all alerts for the supplied deployment_id
}
\description{
Get information of all alerts related to a deployment_id.
}
