% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{get_deployment_api_keys}
\alias{get_deployment_api_keys}
\title{Get API keys for a deployed model.}
\usage{
get_deployment_api_keys(deployment_id)
}
\arguments{
\item{deployment_id}{id of the deployment to get API keys, can be obtained with get_deployments().}
}
\value{
data.frame - API keys available for deployment_id.
}
\description{
Get API keys for a deployed model.
}
