% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{get_deployment_info}
\alias{get_deployment_info}
\title{Get information about a deployment from its id.}
\usage{
get_deployment_info(deployment_id)
}
\arguments{
\item{deployment_id}{id of the deployment to be retrieved, can be obtained with get_deployments().}
}
\value{
list - parsed content of the deployment.
}
\description{
Get information about a deployment from its id.
}
