% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{get_deployment_predictions}
\alias{get_deployment_predictions}
\title{Get listing of predictions related to a deployment_id.}
\usage{
get_deployment_predictions(deployment_id)
}
\arguments{
\item{deployment_id}{id of the deployment, can be obtained with get_deployments().}
}
\value{
list - predictions available for a deployed model.
}
\description{
Get listing of predictions related to a deployment_id.
}
