% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{get_deployment_usage}
\alias{get_deployment_usage}
\title{Get usage (calls, errors and response time) of the last version of a deployed model.}
\usage{
get_deployment_usage(deployment_id, usage_type)
}
\arguments{
\item{deployment_id}{id of the deployment to get usage, can be obtained with get_deployments().}

\item{usage_type}{type of usage to get, among "calls", "errors", "response_time".}
}
\value{
list - plotly object.
}
\description{
Get usage (calls, errors and response time) of the last version of a deployed model.
}
