% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{get_deployments}
\alias{get_deployments}
\title{Get information of all deployments of a given type available for a given project_id.}
\usage{
get_deployments(project_id, type)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{type}{type of the deployment to retrieve among "model" or "app".}
}
\value{
list - parsed content of all deployments of the given type for the supplied project_id.
}
\description{
Get information of all deployments of a given type available for a given project_id.
}
