% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_id_from_name}
\alias{get_experiment_id_from_name}
\title{Get a experiment_id from a experiment_name If duplicated name, the first experiment_id that match it is retrieved.}
\usage{
get_experiment_id_from_name(project_id, experiment_name)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{experiment_name}{name of the experiment we are searching its id from. Can be obtained with get_experiments().}
}
\value{
character - id matching the experiment_name if found.
}
\description{
Get a experiment_id from a experiment_name If duplicated name, the first experiment_id that match it is retrieved.
}
