% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_info}
\alias{get_experiment_info}
\title{Get a experiment from its experiment_id.}
\usage{
get_experiment_info(experiment_id)
}
\arguments{
\item{experiment_id}{id of the experiment, can be obtained with get_experiments().}
}
\value{
list - parsed content of the experiment.
}
\description{
Get a experiment from its experiment_id.
}
