% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_version_features}
\alias{get_experiment_version_features}
\title{Get features information related to a experiment_version_id.}
\usage{
get_experiment_version_features(experiment_version_id)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}
}
\value{
list - parsed content of the experiment_version features information.
}
\description{
Get features information related to a experiment_version_id.
}
