% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_version_info}
\alias{get_experiment_version_info}
\title{Get a experiment_version info from its experiment_version_id}
\usage{
get_experiment_version_info(experiment_version_id)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}
}
\value{
list - parsed content of the experiment_version.
}
\description{
Get a experiment_version info from its experiment_version_id
}
