% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_version_models}
\alias{get_experiment_version_models}
\title{Get a model list related to a experiment_version_id.}
\usage{
get_experiment_version_models(experiment_version_id)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}
}
\value{
list - parsed content of models attached to experiment_version_id.
}
\description{
Get a model list related to a experiment_version_id.
}
