% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment_version_predictions}
\alias{get_experiment_version_predictions}
\title{Get a list of prediction from a experiment_version_id.}
\usage{
get_experiment_version_predictions(
  experiment_version_id,
  generating_type = "user"
)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}

\item{generating_type}{can be "user" (= user predictions) or "auto" (= hold out predictions).}
}
\value{
list - parsed prediction list items.
}
\description{
Get a list of prediction from a experiment_version_id.
}
