% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiments}
\alias{get_experiments}
\title{Get information of all experiments available for a given project_id.}
\usage{
get_experiments(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
list - parsed content of all experiments for the supplied project_id.
}
\description{
Get information of all experiments available for a given project_id.
}
