% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{get_exporter_exports}
\alias{get_exporter_exports}
\title{Get all exports done from an exporter_id}
\usage{
get_exporter_exports(exporter_id)
}
\arguments{
\item{exporter_id}{id of the exporter to retrieve information, can be obtained with get_exporters().}
}
\value{
list - list of exports of the supplied exporter_id.
}
\description{
Get all exports done from an exporter_id
}
