% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{get_exporter_id_from_name}
\alias{get_exporter_id_from_name}
\title{Get a exporter_id from a exporter_name. If duplicated name, the first exporter_id that match it is retrieved}
\usage{
get_exporter_id_from_name(project_id, exporter_name)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{exporter_name}{name of the exporter we are searching its id from. Can be obtained with get_exporters().}
}
\value{
character - id of the exporter if found.
}
\description{
Get a exporter_id from a exporter_name. If duplicated name, the first exporter_id that match it is retrieved
}
