% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{get_exporter_info}
\alias{get_exporter_info}
\title{Get an exporter from its id.}
\usage{
get_exporter_info(exporter_id)
}
\arguments{
\item{exporter_id}{id of the exporter to be retrieved, can be obtained with get_exporters().}
}
\value{
list - parsed content of the exporter.
}
\description{
Get an exporter from its id.
}
