% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{get_exporters}
\alias{get_exporters}
\title{Get information of all exporters available for a given project_id.}
\usage{
get_exporters(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
parsed content of all exporters for the suppled project_id.
}
\description{
Get information of all exporters available for a given project_id.
}
