% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_features_infos}
\alias{get_features_infos}
\title{Get information of a given feature related to a experiment_version_id.}
\usage{
get_features_infos(experiment_version_id, feature_name)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}

\item{feature_name}{name of the feature to retrive information.}
}
\value{
parsed content of the specific feature.
}
\description{
Get information of a given feature related to a experiment_version_id.
}
