% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{get_folder}
\alias{get_folder}
\title{Get a folder from its id.}
\usage{
get_folder(project_id, folder_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{folder_id}{id of the image folder, can be obtained with get_folders().}
}
\value{
parsed content of the folder.
}
\description{
Get a folder from its id.
}
