% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_model_cv}
\alias{get_model_cv}
\title{Get the cross validation file from a specific model.}
\usage{
get_model_cv(model_id)
}
\arguments{
\item{model_id}{id of the model to get the CV, can be obtained with get_usecase_version_models().}
}
\value{
a dataframe containing cross validation data.
}
\description{
Get the cross validation file from a specific model.
}
