% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_model_hyperparameters}
\alias{get_model_hyperparameters}
\title{Get hyperparameters corresponding to a model_id.}
\usage{
get_model_hyperparameters(model_id)
}
\arguments{
\item{model_id}{id of the model, can be obtained with experimentModels(experiment_id).}
}
\value{
list - parsed content of the model's hyperparameters.
}
\description{
Get hyperparameters corresponding to a model_id.
}
