% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_model_infos}
\alias{get_model_infos}
\title{Get model information corresponding to a model_id.}
\usage{
get_model_infos(model_id)
}
\arguments{
\item{model_id}{id of the model, can be obtained with get_experiment_models().}
}
\value{
list - parsed content of the model.
}
\description{
Get model information corresponding to a model_id.
}
