% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_pipeline_id_from_name}
\alias{get_pipeline_id_from_name}
\title{Get a pipeline_id from a pipeline_name and type for a given project_id. If duplicated name, the first pipeline_id that match it is retrieved.}
\usage{
get_pipeline_id_from_name(project_id, name, type)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{name}{name of the pipeline we are searching its id from.}

\item{type}{type of the pipeline to be retrieved among "component", "template", "run".}
}
\value{
character - id of the connector if found.
}
\description{
Get a pipeline_id from a pipeline_name and type for a given project_id. If duplicated name, the first pipeline_id that match it is retrieved.
}
