% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_pipelines}
\alias{get_pipelines}
\title{Get information of all pipelines of a given type available for a given project_id.}
\usage{
get_pipelines(project_id, type)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{type}{type of the pipeline to retrieve among "component", "template", or "run".}
}
\value{
list - parsed content of all pipelines of the given type for the supplied project_id.
}
\description{
Get information of all pipelines of a given type available for a given project_id.
}
