% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_prediction}
\alias{get_prediction}
\title{Get a specific prediction from a prediction_id. Wait up until time_out is reached and wait wait_time between each retry.}
\usage{
get_prediction(prediction_id, prediction_type, time_out = 3600, wait_time = 10)
}
\arguments{
\item{prediction_id}{id of the prediction to be retrieved, can be obtained with get_experiment_version_predictions().}

\item{prediction_type}{type of prediction among "validation" (not deployed model) and "deployment" (deployed model).}

\item{time_out}{maximum number of seconds to wait for the prediction. 3 600 by default.}

\item{wait_time}{number of seconds to wait between each retry. 10 by default.}
}
\value{
data.frame - predictions coming from prediction_id.
}
\description{
Get a specific prediction from a prediction_id. Wait up until time_out is reached and wait wait_time between each retry.
}
