% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_datasets}
\alias{get_project_datasets}
\title{Get datasets from a project.}
\usage{
get_project_datasets(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
parsed content of the project's datasets
}
\description{
Get datasets from a project.
}
