% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project_id_from_name}
\alias{get_project_id_from_name}
\title{Get a project_id from a project_name If duplicated name, the first project_id that match it is retrieved.}
\usage{
get_project_id_from_name(project_name)
}
\arguments{
\item{project_name}{name of the project we are searching its id from. Can be obtained with get_projects().}
}
\value{
project_id of the project_name if found.
}
\description{
Get a project_id from a project_name If duplicated name, the first project_id that match it is retrieved.
}
