% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_info}
\alias{get_project_info}
\title{Get a project from its project_id.}
\usage{
get_project_info(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
parsed content of the project
}
\description{
Get a project from its project_id.
}
