% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project_users}
\alias{get_project_users}
\title{Get users from a project.}
\usage{
get_project_users(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
list - information of project's users.
}
\description{
Get users from a project.
}
