% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{get_scheduler_id_from_name}
\alias{get_scheduler_id_from_name}
\title{Get an scheduler_id from an scheduler_name If duplicated name, the first scheduler_id that match it is retrieved.}
\usage{
get_scheduler_id_from_name(scheduler_name)
}
\arguments{
\item{scheduler_name}{name of the scheduler task we are searching its id from. Can be obtained with get_scheduler().}
}
\value{
id of the scheduler task if found.
}
\description{
Get an scheduler_id from an scheduler_name If duplicated name, the first scheduler_id that match it is retrieved.
}
