% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_shared_usecase_users}
\alias{get_shared_usecase_users}
\title{Get the list of users that can access the usecase.}
\usage{
get_shared_usecase_users(usecase_id)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}
}
\description{
Get the list of users that can access the usecase.
}
