% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_cv}
\alias{get_usecase_cv}
\title{Get the cross validation file from a specific model of a given usecase / version.}
\usage{
get_usecase_cv(usecase_id, model_id, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{model_id}{id of the model to get the CV, can be obtained with usecaseModels(usecase_id).}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\value{
a dataframe containing cross validation data.
}
\description{
Get the cross validation file from a specific model of a given usecase / version.
}
