% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_info}
\alias{get_usecase_info}
\title{Get a usecase from its usecase_id and its version number.}
\usage{
get_usecase_info(usecase_id, version_number = NULL)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{version_number}{number of the version of the usecase. If not supplied, retrieve all usecase information}
}
\value{
parsed content of the usecase.
}
\description{
Get a usecase from its usecase_id and its version number.
}
