% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_predictions}
\alias{get_usecase_predictions}
\title{Get a usecase from its usecase_id and its version number.}
\usage{
get_usecase_predictions(
  usecase_id,
  generating_type = "user",
  version_number = 1
)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{generating_type}{can be "user" (= user predictions) or "auto" (= hold out predictions).}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\value{
parsed prediction list items.
}
\description{
Get a usecase from its usecase_id and its version number.
}
