% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_version_features}
\alias{get_usecase_version_features}
\title{Get features information related to a usecase_version_id.}
\usage{
get_usecase_version_features(usecase_version_id)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}
}
\value{
parsed content of the usecase_version features information.
}
\description{
Get features information related to a usecase_version_id.
}
