% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_version_id}
\alias{get_usecase_version_id}
\title{Get a usecase version id from a usecase_id and its version number.}
\usage{
get_usecase_version_id(usecase_id, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{version_number}{number of the version of the usecase. 1 by default}
}
\value{
usecase version id.
}
\description{
Get a usecase version id from a usecase_id and its version number.
}
