% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_version_models}
\alias{get_usecase_version_models}
\title{Get a model list related to a usecase_version_id.}
\usage{
get_usecase_version_models(usecase_version_id)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}
}
\value{
parsed content of models attached to usecase_version_id.
}
\description{
Get a model list related to a usecase_version_id.
}
