% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_version_predictions}
\alias{get_usecase_version_predictions}
\title{Get a list of prediction from a usecase_version_id.}
\usage{
get_usecase_version_predictions(usecase_version_id, generating_type = "user")
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}

\item{generating_type}{can be "user" (= user predictions) or "auto" (= hold out predictions).}
}
\value{
parsed prediction list items.
}
\description{
Get a list of prediction from a usecase_version_id.
}
