% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecases}
\alias{get_usecases}
\title{Get information of all usecases available for a given project_id.}
\usage{
get_usecases(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
parsed content of all usecases for the suppled project_id.
}
\description{
Get information of all usecases available for a given project_id.
}
