% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_users_usage}
\alias{get_users_usage}
\title{Get Prevision.io users' usage of the given instance}
\usage{
get_users_usage()
}
\value{
Prevision.io users usage
}
\description{
Get Prevision.io users' usage of the given instance
}
