% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_version}
\alias{get_version}
\title{Get Prevision.io version number of the given instance}
\usage{
get_version()
}
\value{
Prevision.io version number
}
\description{
Get Prevision.io version number of the given instance
}
