% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getApp}
\alias{getApp}
\title{Get an app from its id.}
\usage{
getApp(appId)
}
\arguments{
\item{appId}{id of the app to be retrieved, can be obtained with listApps().}
}
\value{
parsed content of the app.
}
\description{
Get an app from its id.
}
