% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getAppIdFromName}
\alias{getAppIdFromName}
\title{Get an appId from an appName If duplicated name, the first appId that match it is retrieved}
\usage{
getAppIdFromName(appName)
}
\arguments{
\item{appName}{name of the app we are searching its id from. Can be obtained with listApps().}
}
\value{
id of the app if found.
}
\description{
Get an appId from an appName If duplicated name, the first appId that match it is retrieved
}
