% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getBestModelId}
\alias{getBestModelId}
\title{Get the modelId that provide the best predictive performance given usecaseId and versionNumber. If includeBlend is false, it will return the modelId from the best "non blended" model.}
\usage{
getBestModelId(usecaseId, versionNumber = 1, includeBlend = TRUE)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}

\item{includeBlend}{boolean, indicating if you want to retrieve the best model among blended models too.}
}
\value{
modelId
}
\description{
Get the modelId that provide the best predictive performance given usecaseId and versionNumber. If includeBlend is false, it will return the modelId from the best "non blended" model.
}
