% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{getConnectorIdFromName}
\alias{getConnectorIdFromName}
\title{Get a connectorId from a connectorName. If duplicated name, the first connectorId that match it is retrieved}
\usage{
getConnectorIdFromName(connectorName)
}
\arguments{
\item{connectorName}{name of the connector we are searching its id from. Can be obtained with getConnectors().}
}
\value{
id of the connector if found.
}
\description{
Get a connectorId from a connectorName. If duplicated name, the first connectorId that match it is retrieved
}
