% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{getConnectorInfos}
\alias{getConnectorInfos}
\title{Get informations about connector from its id.}
\usage{
getConnectorInfos(connectorId)
}
\arguments{
\item{connectorId}{id of the connector to be retrieved, can be obtained with getConnectors().}
}
\value{
parsed content of the connector.
}
\description{
Get informations about connector from its id.
}
