% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{getConnectors}
\alias{getConnectors}
\title{Get informations of all connectors available.}
\usage{
getConnectors()
}
\value{
parsed content of all connectors
}
\description{
Get informations of all connectors available.
}
