% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getDatasetEmbedding}
\alias{getDatasetEmbedding}
\title{Get a dataset embedding from a datasetId.}
\usage{
getDatasetEmbedding(datasetId)
}
\arguments{
\item{datasetId}{dataset id.}
}
\description{
Get a dataset embedding from a datasetId.
}
