% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getDatasetHead}
\alias{getDatasetHead}
\title{Show the head of a dataset from its id.}
\usage{
getDatasetHead(datasetId)
}
\arguments{
\item{datasetId}{id of the dataset, can be obtained with getDatasets().}
}
\value{
head of the dataset as a data.frame object.
}
\description{
Show the head of a dataset from its id.
}
