% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getDatasetIdFromName}
\alias{getDatasetIdFromName}
\title{Get a datasetId from a datasetName If duplicated name, the first datasetId that match it is retrieved}
\usage{
getDatasetIdFromName(datasetName)
}
\arguments{
\item{datasetName}{name of the dataset we are searching its id from. Can be obtained with getDatasets().}
}
\value{
id of the dataset if found.
}
\description{
Get a datasetId from a datasetName If duplicated name, the first datasetId that match it is retrieved
}
