% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getDatasetInfos}
\alias{getDatasetInfos}
\title{Get a dataset from its id.}
\usage{
getDatasetInfos(datasetId)
}
\arguments{
\item{datasetId}{id of the dataset, can be obtained with getDatasets().}
}
\value{
parsed content of the dataset.
}
\description{
Get a dataset from its id.
}
