% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getDatasets}
\alias{getDatasets}
\title{Get informations of all tabular datasets availables.}
\usage{
getDatasets()
}
\value{
parsed content of all datasets.
}
\description{
Get informations of all tabular datasets availables.
}
