% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{getDatasourceIdFromName}
\alias{getDatasourceIdFromName}
\title{Get a datasourceId from a datasourceName If duplicated name, the first datasourceId that match it is retrieved}
\usage{
getDatasourceIdFromName(datasourceName)
}
\arguments{
\item{datasourceName}{name of the connector we are searching its id from. Can be obtained with getDatasources().}
}
\value{
id of the datasource if found.
}
\description{
Get a datasourceId from a datasourceName If duplicated name, the first datasourceId that match it is retrieved
}
