% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{getDatasourceInfos}
\alias{getDatasourceInfos}
\title{Get a datasource from its id.}
\usage{
getDatasourceInfos(datasourceId)
}
\arguments{
\item{datasourceId}{id of the datasources to be retrieved, can be obtained with getDatasources().}
}
\value{
parsed content of the datasources
}
\description{
Get a datasource from its id.
}
