% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{getDatasources}
\alias{getDatasources}
\title{Get informations of all data sources available.}
\usage{
getDatasources()
}
\value{
parsed content of all datasources
}
\description{
Get informations of all data sources available.
}
