% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getFeaturesInfos}
\alias{getFeaturesInfos}
\title{Get information of a given feature related to a usecaseId and its version number}
\usage{
getFeaturesInfos(usecaseId, featureName, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{featureName}{name of the feature to retrive information}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the specific feature
}
\description{
Get information of a given feature related to a usecaseId and its version number
}
