% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{getFolder}
\alias{getFolder}
\title{Get a folder from its id.}
\usage{
getFolder(folder_id)
}
\arguments{
\item{folder_id}{id of the image folder, can be obtained with getFolders().}
}
\value{
parsed content of the dataset.
}
\description{
Get a folder from its id.
}
