% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{getFolders}
\alias{getFolders}
\title{Get informations of all folders availables.}
\usage{
getFolders()
}
\value{
parsed content of all folders.
}
\description{
Get informations of all folders availables.
}
