% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getModelFeatureImportance}
\alias{getModelFeatureImportance}
\title{Get feature importance corresponding to usecaseId and modelId and its version number}
\usage{
getModelFeatureImportance(usecaseId, modelId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{modelId}{id of the model, can be obtained with usecaseModels(usecaseId).}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
dataset of the model's feature importance
}
\description{
Get feature importance corresponding to usecaseId and modelId and its version number
}
