% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getModelHyperparameters}
\alias{getModelHyperparameters}
\title{Get hyperparameters corresponding to usecaseId and modelId and its version number}
\usage{
getModelHyperparameters(usecaseId, modelId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{modelId}{id of the model, can be obtained with usecaseModels(usecaseId).}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the model's hyperparameters
}
\description{
Get hyperparameters corresponding to usecaseId and modelId and its version number
}
