% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getModelInfos}
\alias{getModelInfos}
\title{Get model informations corresponding to usecaseId and modelId and its version number}
\usage{
getModelInfos(usecaseId, modelId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{modelId}{id of the model, can be obtained with usecaseModels(usecaseId).}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the model
}
\description{
Get model informations corresponding to usecaseId and modelId and its version number
}
