% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getPredictionEvents}
\alias{getPredictionEvents}
\title{Get events for prediction of usecaseId and its version number}
\usage{
getPredictionEvents(usacaseId, versionNumber = 1)
}
\arguments{
\item{versionNumber}{number of the version of the usecase. 1 by default.}

\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}
}
\value{
parsed event list
}
\description{
Get events for prediction of usecaseId and its version number
}
