% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getPredictionInfos}
\alias{getPredictionInfos}
\title{Get a informations about a prediction from a specific usecase / version number}
\usage{
getPredictionInfos(usecaseId, predictionId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{predictionId}{id of the prediction to be retrieved, can be obtained with usecasePredictions()}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
list of prediction informations
}
\description{
Get a informations about a prediction from a specific usecase / version number
}
