% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{getSchedulerIdFromName}
\alias{getSchedulerIdFromName}
\title{Get an schedulerId from an schedulerName If duplicated name, the first schedulerId that match it is retrieved}
\usage{
getSchedulerIdFromName(schedulerName)
}
\arguments{
\item{schedulerName}{name of the scheduler task we are searching its id from. Can be obtained with listScheduler().}
}
\value{
id of the scheduler task if found.
}
\description{
Get an schedulerId from an schedulerName If duplicated name, the first schedulerId that match it is retrieved
}
