% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getSharedUsecaseUsers}
\alias{getSharedUsecaseUsers}
\title{Get the list of users that can access the usecase}
\usage{
getSharedUsecaseUsers(usecaseId)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}
}
\description{
Get the list of users that can access the usecase
}
