% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseCV}
\alias{getUsecaseCV}
\title{Get the cross validation file from a specific model of a given usecase / version}
\usage{
getUsecaseCV(usecaseId, modelId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{modelId}{id of the model to get the CV, can be obtained with usecaseModels(usecaseId).}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
a dataframe
}
\description{
Get the cross validation file from a specific model of a given usecase / version
}
