% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseEvents}
\alias{getUsecaseEvents}
\title{Get events of a version of ausecase. If no usecaseId provided, get events from all usecases}
\usage{
getUsecaseEvents(usecaseId = NULL, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed event list
}
\description{
Get events of a version of ausecase. If no usecaseId provided, get events from all usecases
}
