% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseFeatures}
\alias{getUsecaseFeatures}
\title{Get features informations related to a usecaseId and its version number}
\usage{
getUsecaseFeatures(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the usecase features informations
}
\description{
Get features informations related to a usecaseId and its version number
}
