% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseIdFromName}
\alias{getUsecaseIdFromName}
\title{Get a usecaseId from a usecaseName If duplicated name, the first usecaseId that match it is retrieved}
\usage{
getUsecaseIdFromName(usecaseName)
}
\arguments{
\item{usecaseName}{name of the usecase we are searching its id from. Can be obtained with getUsecases().}
}
\value{
usecaseId of the usecaseName if found.
}
\description{
Get a usecaseId from a usecaseName If duplicated name, the first usecaseId that match it is retrieved
}
