% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseModels}
\alias{getUsecaseModels}
\title{Get a model list related to a usecaseId and its version number}
\usage{
getUsecaseModels(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of models attached to usecaseId
}
\description{
Get a model list related to a usecaseId and its version number
}
