% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecasePredictions}
\alias{getUsecasePredictions}
\title{Get a usecase from its usecaseId and its version number}
\usage{
getUsecasePredictions(usecaseId, generatingType = "user", versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{generatingType}{can be "user" (= user predictions) or "auto" (= hold out predictions)}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed prediction list items
}
\description{
Get a usecase from its usecaseId and its version number
}
