% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseSchema}
\alias{getUsecaseSchema}
\title{Get schema related to a usecaseId and its version number}
\usage{
getUsecaseSchema(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the usecase schema
}
\description{
Get schema related to a usecaseId and its version number
}
