% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{getUsecaseTasks}
\alias{getUsecaseTasks}
\title{Get all tasks related to a usecaseId and its version number}
\usage{
getUsecaseTasks(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the usecase tasks
}
\description{
Get all tasks related to a usecaseId and its version number
}
