% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getVersion}
\alias{getVersion}
\title{Get Prevision.io version number of the given instance}
\usage{
getVersion()
}
\value{
Prevision.io version number
}
\description{
Get Prevision.io version number of the given instance
}
