% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_cv_classif_analysis}
\alias{helper_cv_classif_analysis}
\title{Get metrics on a CV file retrieved by Prevision.io for a binary classification use case}
\usage{
helper_cv_classif_analysis(actual, predicted, fold, thresh = NULL, step = 1000)
}
\arguments{
\item{actual}{target comming from the cross Validation dataframe retrieved by Prevision.io}

\item{predicted}{prediction comming from the cross Validation dataframe retrieved by Prevision.io}

\item{fold}{fold number comming from the cross Validation dataframe retrieved by Prevision.io}

\item{thresh}{threshold to use. If not provided optimal threshold given F1 score will be computed}

\item{step}{number of iteration done to find optimal thresh (1000 by default = 0.1\% resolution per fold)}
}
\value{
data.frame with metrics computed on the CV
}
\description{
Get metrics on a CV file retrieved by Prevision.io for a binary classification use case
}
