% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_drift_analysis}
\alias{helper_drift_analysis}
\title{[BETA] Return a data.frame that contains features, a boolean indicating if the feature may have a different distribution between the submitted datasets (if p-value < threshold), their exact p-value and the test used to compute it.}
\usage{
helper_drift_analysis(dataset_1, dataset_2, p_value = 0.05, features = NULL)
}
\arguments{
\item{dataset_1}{the first data set}

\item{dataset_2}{the second data set}

\item{p_value}{a p-value that will be the decision criteria for deciding if a feature is suspicious 5\% by default}

\item{features}{a vector of features names that should be tested. If NULL, only the intersection of the names() will be kept}
}
\value{
vector - a vector of suspicious features.
}
\description{
[BETA] Return a data.frame that contains features, a boolean indicating if the feature may have a different distribution between the submitted datasets (if p-value < threshold), their exact p-value and the test used to compute it.
}
