% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_plot_classif_analysis}
\alias{helper_plot_classif_analysis}
\title{Plot RECALL, PRECISION & F1 SCORE versus top n predictions for a binary classification use case}
\usage{
helper_plot_classif_analysis(actual, predicted, top, compute_every_n = 1)
}
\arguments{
\item{actual}{true value (0 or 1 only)}

\item{predicted}{prediction vector (probability)}

\item{top}{top individual to analyse}

\item{compute_every_n}{compute indicators every n individuals (1 by default)}
}
\value{
data.frame - metrics computed between actual and predicted vectors.
}
\description{
Plot RECALL, PRECISION & F1 SCORE versus top n predictions for a binary classification use case
}
