% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{initPrevisionioClient}
\alias{initPrevisionioClient}
\title{Initialization of the Client on the platform.}
\usage{
initPrevisionioClient(token, url)
}
\arguments{
\item{token}{String, your master token available on the platform.}

\item{url}{String, the url of your access page to the platform.}
}
\description{
Thanks to the token and the url, you can create a Prevision.io Client.
}
\examples{
\dontrun{initPrevisionioClient(token, 'https://xxx.prevision.io')}

}
