% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kc_house}
\alias{kc_house}
\title{House sales in king country}
\format{
A data.frame with 10000 rows and 21 variables
}
\usage{
kc_house
}
\description{
This dataset comes from the website kaggle
<https://www.kaggle.com/harlfoxem/housesalesprediction>
Dataset for regression with target "price"
}
\examples{
kc_house
}
\keyword{datasets}
