% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{optimalPrediction}
\alias{optimalPrediction}
\title{Compute the optimal prediction for each rows in a data frame, for a given model, a list of actionable features and a number of samples for each features to be tested}
\usage{
optimalPrediction(
  usecaseId,
  modelId,
  df,
  actionable,
  nbSample,
  maximize,
  zip = F
)
}
\arguments{
\item{usecaseId}{the id of the usecase to be predicted on}

\item{modelId}{the id of the model to be predicted on}

\item{df}{a data frame to be predicted on}

\item{actionable}{a list of actionables features contained in the names of the data frame}

\item{nbSample}{a vector of number of sample for each actionable features}

\item{maximize}{a boolean indicating if we maximize or minimize the predicted target}

\item{zip}{a boolean indicating if the data frame to predict should be zipped prior sending to the instance}
}
\value{
row data.frame with the optimal vector and the prediction associated with for each rows in the original data frame
}
\description{
Compute the optimal prediction for each rows in a data frame, for a given model, a list of actionable features and a number of samples for each features to be tested
}
