% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{pause_usecase_version}
\alias{pause_usecase_version}
\title{Pause a running usecase_version on the platform.}
\usage{
pause_usecase_version(usecase_version_id)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}
}
\description{
Pause a running usecase_version on the platform.
}
