% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{pauseUsecase}
\alias{pauseUsecase}
\title{Pause a version of a usecase on the platform}
\usage{
pauseUsecase(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\description{
Pause a version of a usecase on the platform
}
