% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pio_download}
\alias{pio_download}
\title{Download resources according specific parameters.}
\usage{
pio_download(endpoint, tempFile)
}
\arguments{
\item{endpoint}{end of the url of the API call.}

\item{tempFile}{temporary file to download.}
}
\description{
Download resources according specific parameters.
}
