% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pio_init}
\alias{pio_init}
\title{Initialization of the connection to your instance Prevision.io.}
\usage{
pio_init(token, url)
}
\arguments{
\item{token}{your master token, can be found on your instance on the "API KEY" page.}

\item{url}{the url of your instance.}
}
\value{
list - url and token needed for connecting to the Prevision.io environment.
}
\description{
Initialization of the connection to your instance Prevision.io.
}
\examples{
\dontrun{pio_init('eyJhbGciOiJIUz', 'https://xxx.prevision.io')}

}
