% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pio_list_to_df}
\alias{pio_list_to_df}
\title{Convert a list returned from APIs to a dataframe. Only working for consistent list (same naming and number of columns).}
\usage{
pio_list_to_df(list)
}
\arguments{
\item{list}{named list comming from an API call.}
}
\value{
data.frame - cast a consistent list to a data.frame.
}
\description{
Convert a list returned from APIs to a dataframe. Only working for consistent list (same naming and number of columns).
}
