% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pio_request}
\alias{pio_request}
\title{Request the platform. Thanks to an endpoint, the url and the API, you can create request.}
\usage{
pio_request(endpoint, method, data = NULL, upload = FALSE)
}
\arguments{
\item{endpoint}{end of the url of the API call.}

\item{method}{the method needed according the API (Available: POST, GET, DELETE).}

\item{data}{object to upload when using method POST.}

\item{upload}{used parameter when uploading dataset (for encoding in API call), don't use it.}
}
\value{
list - response from the request.
}
\description{
Request the platform. Thanks to an endpoint, the url and the API, you can create request.
}
\examples{
\dontrun{pio_request(paste0('/jobs/', experiment$jobId), DELETE)}

}
