% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotClassifAnalysis}
\alias{plotClassifAnalysis}
\title{Plot RECALL, PRECISION & F1 SCORE versus top n predictions for a binary classification use case}
\usage{
plotClassifAnalysis(actual, predicted, top)
}
\arguments{
\item{actual}{true value (0 or 1 only)}

\item{predicted}{prediction vector (probability)}

\item{top}{top individual to analyse}
}
\value{
data.frame with metrics computed on the CV
}
\description{
Plot RECALL, PRECISION & F1 SCORE versus top n predictions for a binary classification use case
}
