% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{previsionDownload}
\alias{previsionDownload}
\title{Download according specific parameters}
\usage{
previsionDownload(endpoint, tempFile)
}
\arguments{
\item{endpoint}{string, end of the url of the API call.}

\item{tempFile}{file, tempory file to be download.}
}
\value{
write on disk the content of the temporary file.
}
\description{
Download according specific parameters
}
