% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{previsionioRequest}
\alias{previsionioRequest}
\title{Request the platform}
\usage{
previsionioRequest(endpoint, method, data = NULL, upload = FALSE)
}
\arguments{
\item{endpoint}{string, end of the url of the API call.}

\item{method}{string, the method needed according the API (Available: POST, GET, DELETE).}

\item{data, }{oject to upload when using method POST.}

\item{upload, }{used parameter when uploading dataset (for encoding in API call), don't use it.}
}
\description{
Thanks to an endpoint, the url and the API, you can create request.
}
\examples{
\dontrun{previsionioRequest(paste0('/jobs/', usecase$jobId), DELETE)}

}
