% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{resume_experiment_version}
\alias{resume_experiment_version}
\title{Resume a paused experiment_version on the platform.}
\usage{
resume_experiment_version(experiment_version_id)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}
}
\description{
Resume a paused experiment_version on the platform.
}
