% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{resume_usecase}
\alias{resume_usecase}
\title{Resume a version of usecase on the platform.}
\usage{
resume_usecase(usecase_id, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\description{
Resume a version of usecase on the platform.
}
