% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{resumeUsecase}
\alias{resumeUsecase}
\title{Resume a version of usecase on the platform}
\usage{
resumeUsecase(usecaseId, versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\description{
Resume a version of usecase on the platform
}
