% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{run_scheduler}
\alias{run_scheduler}
\title{Run a scheduled task.}
\usage{
run_scheduler(scheduler_id)
}
\arguments{
\item{scheduler_id}{id of the scheduler task to be run. Can be obtained with get_scheduler().}
}
\value{
response status code (200 on success)
}
\description{
Run a scheduled task.
}
