% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{share_usecase}
\alias{share_usecase}
\title{Share a usecase to a specific user}
\usage{
share_usecase(usecase_id, email)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{email}{email adress of the user you want to share the usecase with. Should be a platform user.}
}
\value{
list of shared users.
}
\description{
Share a usecase to a specific user
}
