% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{shareUsecase}
\alias{shareUsecase}
\title{Share a usecase to a specific user}
\usage{
shareUsecase(usecaseId, email)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{email}{email adress of the user you want to share the usecase with. Should be a platform user.}
}
\value{
list of shared users
}
\description{
Share a usecase to a specific user
}
