% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{startDatasetEmbedding}
\alias{startDatasetEmbedding}
\title{Start a dataset embedding from a datasetId.}
\usage{
startDatasetEmbedding(datasetId)
}
\arguments{
\item{datasetId}{dataset id.}
}
\description{
Start a dataset embedding from a datasetId.
}
