% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{startPrediction}
\alias{startPrediction}
\title{Start a prediction on a existing usecase}
\usage{
startPrediction(
  usecaseId,
  datasetId = NULL,
  datasetFolderId = NULL,
  confidence = F,
  bestSingle = F,
  modelId = NULL,
  versionNumber = 1,
  queriesDatasetId = NULL,
  queriesDatasetContentColumn = NULL,
  queriesDatasetIdColumn = NULL,
  queriesDatasetMatchingIdDescriptionColumn = NULL,
  topK = NULL
)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{datasetId}{id of the dataset to start the prediction on, can be obtained with getUsecases()}

\item{datasetFolderId}{id of the folder dataset to start prediction on, can be obtained with getUsecases(). Only usefull for images use cases}

\item{confidence}{boolean. If enable, confidence intervalle will be added to predictions}

\item{bestSingle}{boolean. If enable, best single model (non blend) will be used for making predictions other wise, best model will be used unless if modelId is fed}

\item{modelId}{id of the model to start the prediction on. If provided, it will overwrite the "best single" params}

\item{versionNumber}{number of the version of the usecase. 1 by default}

\item{queriesDatasetId}{id of the datasat containing queries (text-similarity)}

\item{queriesDatasetContentColumn}{name of the content column in the queries dataset (text-similarity)}

\item{queriesDatasetIdColumn}{name of the id columln in the queries dataset (text-similarity)}

\item{queriesDatasetMatchingIdDescriptionColumn}{name of the column matching the id (text-similarity)}

\item{topK}{number of class to retrieve (text-similarity)}
}
\value{
parsed prediction list
}
\description{
Start a prediction on a existing usecase
}
