% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{startScheduler}
\alias{startScheduler}
\title{Start a scheduled task}
\usage{
startScheduler(schedulerId)
}
\arguments{
\item{schedulerId}{id of the scheduler task to be run. Can be obtained with listScheduler().}
}
\value{
response status code (200 on success)
}
\description{
Start a scheduled task
}
