% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{startUsecase}
\alias{startUsecase}
\title{Start a new usecase on the platform.}
\usage{
startUsecase(
  name,
  dataType,
  trainingType,
  datasetId,
  targetColumn = NULL,
  holdoutDatasetId = NULL,
  idColumn = NULL,
  dropList = NULL,
  profile = NULL,
  usecaseDescription = NULL,
  metric = NULL,
  foldColumn = NULL,
  normalModels = NULL,
  liteModels = c("XGB"),
  simpleModels = NULL,
  withBlend = NULL,
  weightColumn = NULL,
  featuresEngineeringSelectedList = NULL,
  featuresSelectionCount = NULL,
  featuresSelectionTime = NULL,
  datasetFolderId = NULL,
  filenameColumn = NULL,
  ymin = NULL,
  ymax = NULL,
  xmin = NULL,
  xmax = NULL,
  timeColumn = NULL,
  startDW = NULL,
  endDW = NULL,
  startFW = NULL,
  endFW = NULL,
  groupList = NULL,
  aprioriList = NULL,
  contentColumn = NULL,
  queriesDatasetId = NULL,
  queriesDatasetContentColumn = NULL,
  queriesDatasetIdColumn = NULL,
  queriesDatasetMatchingIdDescriptionColumn = NULL,
  topK = NULL,
  lang = NULL,
  modelsParameters = NULL
)
}
\arguments{
\item{name}{name of the usecase}

\item{dataType}{type of data ("tabular" or "images" or "timeseries")}

\item{trainingType}{type of the training you want to achieve ("regression", "classification", "multiclassification", "clustering", "object-detection", "text-similarity")}

\item{datasetId}{id of the dataset used for the training phase}

\item{targetColumn}{name of the TARGET column}

\item{holdoutDatasetId}{id of the holdout dataset}

\item{idColumn}{name of the id column}

\item{dropList}{list of names of features to drop}

\item{profile}{chosen profil among "quick", "normal", "advanced"}

\item{usecaseDescription}{usecase description}

\item{metric}{name of the metric to optimise}

\item{foldColumn}{name of the fold column}

\item{normalModels}{list of (normal) models to select with full FE & hyperparameters search}

\item{liteModels}{list of (lite) models to select with lite FE & default hyperparameters}

\item{simpleModels}{list of simple models to select}

\item{withBlend}{do we allow to include blend in the modelisation}

\item{weightColumn}{name of the weight columns}

\item{featuresEngineeringSelectedList}{list of feature engineering to select among "Counter", "Date", "freq", "text_tfidf", "text_word2vec", "text_embedding", "tenc", "poly", "pca", "kmean"}

\item{featuresSelectionCount}{number of features to keep after the feature selection process}

\item{featuresSelectionTime}{time budget in minutes of the feature selection process}

\item{datasetFolderId}{id of the dataset fold (images)}

\item{filenameColumn}{name of the file name path (images)}

\item{ymin}{name of the column matching the lower y value of the image (object detection)}

\item{ymax}{name of the column matching the higher y value of the image (object detection)}

\item{xmin}{name of the column matching the lower x value of the image (object detection)}

\item{xmax}{name of the column matching the higher x value of the image (object detection)}

\item{timeColumn}{name of column containing the timestamp (time series)}

\item{startDW}{value of the start of derivative window (time series), should be a strict negative integer}

\item{endDW}{value of the end of derivative window (time series), should be a negative integer greater than startDW}

\item{startFW}{value of the start of forecast window (time series), should be a strict positive integer}

\item{endFW}{value of the end of forecast window (time series), should be a strict positive integer greater than startFW}

\item{groupList}{list of name of feature that describes groups (time series)}

\item{aprioriList}{list of name of feature that are a priori (time series)}

\item{contentColumn}{content column name (text-similarity)}

\item{queriesDatasetId}{id of the dataset containing queries (text-similarity)}

\item{queriesDatasetContentColumn}{name of the column containing queries in the query dataset (text-similarity)}

\item{queriesDatasetIdColumn}{name of the ID column in the query dataset (text-similarity)}

\item{queriesDatasetMatchingIdDescriptionColumn}{name of the column matching id in the description dataset (text-similarity)}

\item{topK}{top k individual to find (text-similarity)}

\item{lang}{lang of the text (text-similarity)}

\item{modelsParameters}{parameters of the model (text-similarity)}
}
\value{
usecase infos
}
\description{
Start a new usecase on the platform.
}
