% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{stop_experiment_version}
\alias{stop_experiment_version}
\title{Stop a running or paused experiment_version on the platform.}
\usage{
stop_experiment_version(experiment_version_id)
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}
}
\value{
integer - 200 on success.
}
\description{
Stop a running or paused experiment_version on the platform.
}
