% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{stop_usecase}
\alias{stop_usecase}
\title{Stop a version of a usecase on the platform.}
\usage{
stop_usecase(usecase_id, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\description{
Stop a version of a usecase on the platform.
}
