% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{stop_usecase_version}
\alias{stop_usecase_version}
\title{Stop a running or paused usecase_version on the platform.}
\usage{
stop_usecase_version(usecase_version_id)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}
}
\description{
Stop a running or paused usecase_version on the platform.
}
