% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stores_test}
\alias{stores_test}
\title{Stores sales test}
\format{
A data.frame with 118 rows and 6 variables
}
\usage{
stores_test
}
\description{
Dataset for time series with target "Sales" and group variable "Store"
}
\examples{
stores_test
}
\keyword{datasets}
