% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stores_train}
\alias{stores_train}
\title{Stores sales train}
\format{
A data.frame with 1766 rows and 6 variables
}
\usage{
stores_train
}
\description{
Dataset for time series with target "Sales" and group variable "Store"
}
\examples{
stores_train
}
\keyword{datasets}
