% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telecom}
\alias{telecom}
\title{Telecom}
\format{
A data.frame with 7032 rows and 22 variables
}
\usage{
telecom
}
\description{
This dataset comes from the website kaggle
<https://www.kaggle.com/blastchar/telco-customer-churn>
Dataset for clustering
}
\examples{
telecom
}
\keyword{datasets}
