% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{test_connector}
\alias{test_connector}
\title{Test an existing connector.}
\usage{
test_connector(connector_id)
}
\arguments{
\item{connector_id}{id of the connector to be tested, can be obtained with get_connectors().}
}
\description{
Test an existing connector.
}
