% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{test_contact_point}
\alias{test_contact_point}
\title{Test an existing contact point}
\usage{
test_contact_point(contact_point_id)
}
\arguments{
\item{contact_point_id}{id of the contact point to be tested, can be obtained with get_contact_points().}
}
\value{
integer - 200 on success.
}
\description{
Test an existing contact point
}
