% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{test_datasource}
\alias{test_datasource}
\title{Test a datasource}
\usage{
test_datasource(datasource_id)
}
\arguments{
\item{datasource_id}{id of the datasource to be tested, can be obtained with get_datasources().}
}
\description{
Test a datasource
}
