% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{test_deployment_type}
\alias{test_deployment_type}
\title{Check if a type of a deployment is supported}
\usage{
test_deployment_type(type)
}
\arguments{
\item{type}{type of the deployment among "model" or "app".}
}
\value{
no return value, called for side effects.
}
\description{
Check if a type of a deployment is supported
}
