% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{test_pipeline_type}
\alias{test_pipeline_type}
\title{Check if a type of a pipeline is supported}
\usage{
test_pipeline_type(type)
}
\arguments{
\item{type}{type of the pipeline among "component", "template", "run".}
}
\description{
Check if a type of a pipeline is supported
}
