% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{testConnector}
\alias{testConnector}
\title{Test a connector.}
\usage{
testConnector(connectorId)
}
\arguments{
\item{connectorId}{id of the connector to be tested, can be obtained with getConnectors().}
}
\description{
Test a connector.
}
