% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{testDatasource}
\alias{testDatasource}
\title{Test a datasource}
\usage{
testDatasource(datasourceId)
}
\arguments{
\item{datasourceId}{id of the datasource to be tested, can be obtained with getDatasources().}
}
\description{
Test a datasource
}
