% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic}
\format{
A data.frame with 891 rows and 6 variables
}
\usage{
titanic
}
\description{
This dataset comes from the website kaggle <https://www.kaggle.com/c/titanic/data>
Dataset for classification with target "Survived"
}
\examples{
titanic
}
\keyword{datasets}
