% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{unshare_usecase}
\alias{unshare_usecase}
\title{Unshare a use case for the specified email adress. If missing, the usecase will be unshared from all users.}
\usage{
unshare_usecase(usecase_id, email = NULL)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{email}{email adress of the user you want to unshare the usecase with. Should be an instance user.}
}
\value{
list of shared users.
}
\description{
Unshare a use case for the specified email adress. If missing, the usecase will be unshared from all users.
}
