% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{unshareUsecase}
\alias{unshareUsecase}
\title{Unshare a use case for the specified email adress. If missing, the usecase will be unshared from all users}
\usage{
unshareUsecase(usecaseId, email = NULL)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{email}{email adress of the user you want to unshare the usecase with. Should be an instance user.}
}
\value{
list of shared users
}
\description{
Unshare a use case for the specified email adress. If missing, the usecase will be unshared from all users
}
