% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{update_experiment_version_description}
\alias{update_experiment_version_description}
\title{Update the description of a given experiment_version_id.}
\usage{
update_experiment_version_description(experiment_version_id, description = "")
}
\arguments{
\item{experiment_version_id}{id of the experiment_version, can be obtained with get_experiment_version_id().}

\item{description}{Description of the experiment.}
}
\value{
integer - 200 on success.
}
\description{
Update the description of a given experiment_version_id.
}
