% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{update_project_user_role}
\alias{update_project_user_role}
\title{Update user role in and existing project.}
\usage{
update_project_user_role(project_id, user_mail, user_role)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{user_mail}{email of the user to be edit, can be obtained with get_users().}

\item{user_role}{role to grand to the user among "admin", "contributor" and "viewer"}
}
\value{
list of users in the project
}
\description{
Update user role in and existing project.
}
