% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{update_usecase_version_description}
\alias{update_usecase_version_description}
\title{Update the description of a given usecase_version_id.}
\usage{
update_usecase_version_description(usecase_version_id, description = "")
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}

\item{description}{Description of the usecase.}
}
\description{
Update the description of a given usecase_version_id.
}
