% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{updateAppName}
\alias{updateAppName}
\title{Update an app name.}
\usage{
updateAppName(appId, appName)
}
\arguments{
\item{appId}{id of the app to be updated, can be obtained with listApps().}

\item{appName}{new name of the app to be set.}
}
\value{
response status code (200 on success)
}
\description{
Update an app name.
}
