% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{updateusecase_description}
\alias{updateusecase_description}
\title{Update the description of a given usecase and its version number.}
\usage{
updateusecase_description(usecase_id, description = "", version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{description}{Description of the usecase.}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\description{
Update the description of a given usecase and its version number.
}
