% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{updateUsecaseDescription}
\alias{updateUsecaseDescription}
\title{Update the description of a given usecase and its version number}
\usage{
updateUsecaseDescription(usecaseId, description = "", versionNumber = 1)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{description}{Description of the usecase.}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\description{
Update the description of a given usecase and its version number
}
