% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wines}
\alias{wines}
\title{Wines quality}
\format{
A data.frame with 1483 rows and 12 variables
}
\usage{
wines
}
\description{
This dataset comes from the website kaggle
<https://www.kaggle.com/piyushgoyal443/red-wine-dataset>
Dataset for multiclassification with target "quality"
}
\examples{
wines
}
\keyword{datasets}
