% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{create_connector}
\alias{create_connector}
\title{Create a new connector of a supported type (among: "SQL", "FTP", "SFTP", "S3", "GCP").}
\usage{
create_connector(
  project_id,
  type,
  name,
  host,
  port,
  username,
  password,
  google_credentials = NULL
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{type}{connector type.}

\item{name}{connector name.}

\item{host}{connector host.}

\item{port}{connector port.}

\item{username}{connector username.}

\item{password}{connector password.}

\item{google_credentials}{google credentials JSON (for GCP only).}
}
\value{
parsed content of the connector.
}
\description{
Create a new connector of a supported type (among: "SQL", "FTP", "SFTP", "S3", "GCP").
}
