% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataframe_from_dataset}
\alias{create_dataframe_from_dataset}
\title{Create a dataframe from a dataset_id.}
\usage{
create_dataframe_from_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{dataset id.}
}
\value{
data.frame - a R dataframe matching the dataset.
}
\description{
Create a dataframe from a dataset_id.
}
