% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset_from_dataframe}
\alias{create_dataset_from_dataframe}
\title{Upload dataset from data frame.}
\usage{
create_dataset_from_dataframe(project_id, dataset_name, dataframe, zip = FALSE)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{dataset_name}{given name of the dataset on the platform.}

\item{dataframe}{data.frame to upload.}

\item{zip}{is the temp file zipped before sending it to Prevision.io (default = FALSE).}
}
\value{
list - parsed content of the dataset.
}
\description{
Upload dataset from data frame.
}
