% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{create_datasource}
\alias{create_datasource}
\title{Create a new datasource
If check_if_exist is enabled, the function will check if a datasource with the same name already exists. If yes, it will return a message and the information of the existing datasource instead of creating a new one.}
\usage{
create_datasource(
  project_id,
  connector_id,
  name,
  path = "",
  database = "",
  table = "",
  bucket = "",
  request = "",
  check_if_exist = FALSE
)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{connector_id}{connector_id linked to the datasource.}

\item{name}{datasource name.}

\item{path}{datasource path (for SFTP & FTP connector).}

\item{database}{datasource database (for SQL connector).}

\item{table}{datasource table (for SQL connector).}

\item{bucket}{datasource bucket (for S3 connector).}

\item{request}{datasource request (for SQLconnector).}

\item{check_if_exist}{boolean (FALSE by default). If TRUE, makes extra checks to see if a datasource with the same name is already existing.}
}
\value{
list - parsed content of the datasource.
}
\description{
Create a new datasource
If check_if_exist is enabled, the function will check if a datasource with the same name already exists. If yes, it will return a message and the information of the existing datasource instead of creating a new one.
}
