% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{create_folder}
\alias{create_folder}
\title{Upload folder from a local file.}
\usage{
create_folder(project_id, folder_name, file)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}

\item{folder_name}{given name of the folder on the platform.}

\item{file}{path to the folder.}
}
\value{
parsed content of the folder.
}
\description{
Upload folder from a local file.
}
