% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{delete_datasource}
\alias{delete_datasource}
\title{Delete a datasource}
\usage{
delete_datasource(datasource_id)
}
\arguments{
\item{datasource_id}{id of the datasource to be deleted, can be obtained with get_datasources().}
}
\value{
integer - 200 on success.
}
\description{
Delete a datasource
}
