% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{delete_pipeline}
\alias{delete_pipeline}
\title{Delete an existing pipeline}
\usage{
delete_pipeline(pipeline_id, type)
}
\arguments{
\item{pipeline_id}{id of the pipeline to be retrieved, can be obtained with get_pipelines().}

\item{type}{type of the pipeline to be retrieved among "component", "template", "run".}
}
\value{
integer - 204 on success.
}
\description{
Delete an existing pipeline
}
