% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{delete_prediction}
\alias{delete_prediction}
\title{Delete a prediction.}
\usage{
delete_prediction(prediction_id)
}
\arguments{
\item{prediction_id}{id of the prediction to be deleted, can be obtained with get_experiment_version_predictions()}
}
\value{
list of predictions of experiment_id.
}
\description{
Delete a prediction.
}
