% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_best_model_id}
\alias{get_best_model_id}
\title{Get the model_id that provide the best predictive performance given usecase_version_id. If include_blend is false, it will return the model_id from the best "non blended" model.}
\usage{
get_best_model_id(usecase_version_id, include_blend = TRUE)
}
\arguments{
\item{usecase_version_id}{id of the usecase_version, can be obtained with get_usecase_version_id().}

\item{include_blend}{boolean, indicating if you want to retrieve the best model among blended models too.}
}
\value{
model_id.
}
\description{
Get the model_id that provide the best predictive performance given usecase_version_id. If include_blend is false, it will return the model_id from the best "non blended" model.
}
