% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{get_connector_id_from_name}
\alias{get_connector_id_from_name}
\title{Get a connector_id from a connector_name. If duplicated name, the first connector_id that match it is retrieved.}
\usage{
get_connector_id_from_name(connector_name)
}
\arguments{
\item{connector_name}{name of the connector we are searching its id from. Can be obtained with get_connectors().}
}
\value{
id of the connector if found.
}
\description{
Get a connector_id from a connector_name. If duplicated name, the first connector_id that match it is retrieved.
}
