% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{get_connectors}
\alias{get_connectors}
\title{Get information of all connectors available for a given project_id.}
\usage{
get_connectors(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\value{
list - parsed content of all connectors for the supplied project_id.
}
\description{
Get information of all connectors available for a given project_id.
}
