% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_datasets}
\alias{get_datasets}
\title{Get information of all datasets available for a given project_id.}
\usage{
get_datasets(project_id)
}
\arguments{
\item{project_id}{id of the project, can be obtained with get_projects().}
}
\description{
Get information of all datasets available for a given project_id.
}
