% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{get_datasources}
\alias{get_datasources}
\title{Get information of all data sources available.}
\usage{
get_datasources()
}
\value{
parsed content of all data_sources
}
\description{
Get information of all data sources available.
}
