% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_features_infos}
\alias{get_features_infos}
\title{Get information of a given feature related to a usecase_id and its version number.}
\usage{
get_features_infos(usecase_id, feature_name, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{feature_name}{name of the feature to retrive information.}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the specific feature.
}
\description{
Get information of a given feature related to a usecase_id and its version number.
}
