% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{get_folders}
\alias{get_folders}
\title{Get information of all folders availables.}
\usage{
get_folders()
}
\value{
parsed content of all folders.
}
\description{
Get information of all folders availables.
}
