% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_model_infos}
\alias{get_model_infos}
\title{Get model information corresponding to usecase_id and model_id and its version number}
\usage{
get_model_infos(usecase_id, model_id, version_number = 1)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{model_id}{id of the model, can be obtained with usecaseModels(usecase_id).}

\item{version_number}{number of the version of the usecase. 1 by default.}
}
\value{
parsed content of the model.
}
\description{
Get model information corresponding to usecase_id and model_id and its version number
}
