% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_prediction}
\alias{get_prediction}
\title{Get a specific prediction from a specific usecase / version number. Wait up until time_out is reached and wait wait_time between each retry.}
\usage{
get_prediction(
  usecase_id,
  prediction_id,
  version_number = 1,
  time_out = 3600,
  wait_time = 10
)
}
\arguments{
\item{usecase_id}{id of the usecase, can be obtained with get_usecases().}

\item{prediction_id}{id of the prediction to be retrieved, can be obtained with usecasePredictions().}

\item{version_number}{number of the version of the usecase. 1 by default.}

\item{time_out}{maximum number of seconds to wait for the prediction. 3 600 by default.}

\item{wait_time}{number of seconds to wait between each retry. 10 by default.}
}
\value{
a data.frame with the predictions.
}
\description{
Get a specific prediction from a specific usecase / version number. Wait up until time_out is reached and wait wait_time between each retry.
}
