% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_prediction_infos}
\alias{get_prediction_infos}
\title{Get a information about a prediction_id.}
\usage{
get_prediction_infos(prediction_id)
}
\arguments{
\item{prediction_id}{id of the prediction to be retrieved, can be obtained with get_experiment_version_predictions().}
}
\value{
list - parsed prediction information.
}
\description{
Get a information about a prediction_id.
}
