% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{get_usecase_id_from_name}
\alias{get_usecase_id_from_name}
\title{Get a usecase_id from a usecase_name If duplicated name, the first usecase_id that match it is retrieved.}
\usage{
get_usecase_id_from_name(usecase_name)
}
\arguments{
\item{usecase_name}{name of the usecase we are searching its id from. Can be obtained with get_usecases().}
}
\value{
usecase_id of the usecase_name if found.
}
\description{
Get a usecase_id from a usecase_name If duplicated name, the first usecase_id that match it is retrieved.
}
