% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{startPrediction}
\alias{startPrediction}
\title{Start a prediction on a existing usecase with a given datasetId and a given version number}
\usage{
startPrediction(
  usecaseId,
  datasetId,
  datasetFolderId = NULL,
  confidence = F,
  bestSingle = F,
  modelId = NULL,
  versionNumber = 1
)
}
\arguments{
\item{usecaseId}{id of the usecase, can be obtained with getUsecases().}

\item{datasetId}{id of the dataset to start the prediction on, can be obtained with getUsecases()}

\item{datasetFolderId}{id of the folder dataset to start prediction on, can be obtained with getUsecases(). Only usefull for images use cases}

\item{confidence}{boolean. If enable, confidence intervalle will be added to predictions}

\item{bestSingle}{boolean. If enable, best single model (non blend) will be used for making predictions other wise, best model will be used unless if modelId is fed}

\item{modelId}{id of the model to start the prediction on. If provided, it will overwrite the "best single" params}

\item{versionNumber}{number of the version of the usecase. 1 by default.}
}
\value{
parsed prediction list
}
\description{
Start a prediction on a existing usecase with a given datasetId and a given version number
}
