% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecase.R
\name{startUsecase}
\alias{startUsecase}
\title{Start a new usecase on the platform.}
\usage{
startUsecase(
  name,
  dataType,
  trainingType,
  datasetId,
  targetColumn = NULL,
  holdoutDatasetId = NULL,
  idColumn = NULL,
  dropList = NULL,
  profile = NULL,
  metric = NULL,
  foldColumn = NULL,
  normalModels = NULL,
  liteModels = c("XGB"),
  simpleModels = NULL,
  withBlend = NULL,
  weightColumn = NULL,
  featuresEngineeringSelectedList = NULL,
  featuresSelectionCount = NULL,
  featuresSelectionTime = NULL,
  datasetFolderId = NULL,
  filenameColumn = NULL,
  topColumn = NULL,
  bottomColumn = NULL,
  leftColumn = NULL,
  rightColumn = NULL,
  timeColumn = NULL,
  startDW = NULL,
  endDW = NULL,
  startFW = NULL,
  endFW = NULL,
  groupList = NULL,
  aprioriList = NULL
)
}
\arguments{
\item{name}{name of the usecase.}

\item{dataType}{type of data ("tabular" or "images" or "timeseries").}

\item{trainingType}{type of the training you want to achieve ("regression", "classification", "multiclassification", "clustering", "object-detection").}

\item{datasetId}{id of the dataset used for the training phase.}

\item{targetColumn}{name of the TARGET column}

\item{holdoutDatasetId}{id of the houldout dataset}

\item{idColumn}{name of the id column}

\item{dropList}{list of names of features to drop}

\item{profile}{chosen profil among "quick", "normal", "advanced"}

\item{metric}{name of the metric to optimise}

\item{foldColumn}{name of the fold column}

\item{normalModels}{list of (normal) models to select with full FE & hyperparameters search}

\item{liteModels}{list of (lite) models to select with lite FE & default hyperparameters}

\item{simpleModels}{list of simple models to select}

\item{withBlend}{do we allow to include blend in the modelisation}

\item{weightColumn}{name of the weight columns}

\item{featuresEngineeringSelectedList}{list of feature engineering to select among "Counter", "Date", "freq", "text_tfidf", "text_word2vec", "text_embedding", "tenc", "poly", "pca", "kmean"}

\item{featuresSelectionCount}{number of features to keep after the feature selection process}

\item{featuresSelectionTime}{time budget in minutes of the feature selection process}

\item{datasetFolderId}{id of the dataset fold (images)}

\item{filenameColumn}{name of the file name path (images)}

\item{topColumn}{name of the top column (object detection)}

\item{bottomColumn}{name of the bottom column (object detection)}

\item{leftColumn}{name of the left column (object detection)}

\item{rightColumn}{name of the right column (object detection)}

\item{timeColumn}{name of the time column (time series)}

\item{startDW}{value of the start of derivative window (time series), should be a strict negative integer}

\item{endDW}{value of the end of derivative window (time series), should be a negative integer greater than startDW}

\item{startFW}{value of the start of forecast window (time series), should be a strict positive integer}

\item{endFW}{value of the end of forecast window (time series), should be a strict positive integer greater than startFW}

\item{groupList}{list of name of feature that describes groups (time series)}

\item{aprioriList}{list of name of feature that are a priori (time series)}
}
\value{
usecase infos
}
\description{
Start a new usecase on the platform.
}
